# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Fire Tournament API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Player {
  id: number;
  username: string;
  email: string;
  gameId: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface RegisterPlayerBody {
  username: string;
  email: string;
  password: string;
  gameId: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  player: Player;
}

export type MatchType = (typeof MatchType)[keyof typeof MatchType];

export const MatchType = {
  solo: "solo",
  duo: "duo",
  squad: "squad",
} as const;

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;

export interface Match {
  id: number;
  title: string;
  type: MatchType;
  status: MatchStatus;
  scheduledAt: string;
  entryFee: number;
  maxPlayers: number;
  registeredCount: number;
  roomId?: string | null;
  roomPassword?: string | null;
  prizePool?: string | null;
  description?: string | null;
  createdAt: string;
}

export type RegistrationPaymentStatus =
  (typeof RegistrationPaymentStatus)[keyof typeof RegistrationPaymentStatus];

export const RegistrationPaymentStatus = {
  pending: "pending",
  confirmed: "confirmed",
} as const;

export interface Registration {
  id: number;
  matchId: number;
  playerId: number;
  playerUsername: string;
  playerGameId: string;
  teamName?: string | null;
  paymentStatus: RegistrationPaymentStatus;
  registeredAt: string;
}

export type MatchDetail = Match & {
  registrations: Registration[];
};

export type CreateMatchBodyType =
  (typeof CreateMatchBodyType)[keyof typeof CreateMatchBodyType];

export const CreateMatchBodyType = {
  solo: "solo",
  duo: "duo",
  squad: "squad",
} as const;

export interface CreateMatchBody {
  title: string;
  type: CreateMatchBodyType;
  scheduledAt: string;
  maxPlayers: number;
  roomId?: string | null;
  roomPassword?: string | null;
  prizePool?: string | null;
  description?: string | null;
}

export type UpdateMatchBodyStatus =
  (typeof UpdateMatchBodyStatus)[keyof typeof UpdateMatchBodyStatus];

export const UpdateMatchBodyStatus = {
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;

export interface UpdateMatchBody {
  title?: string;
  status?: UpdateMatchBodyStatus;
  scheduledAt?: string;
  roomId?: string | null;
  roomPassword?: string | null;
  prizePool?: string | null;
  description?: string | null;
}

export interface RegisterForMatchBody {
  teamName?: string | null;
}

export type RegistrationWithMatch = Registration & {
  match: Match;
};

export type AdminRegistration = Registration & {
  match: Match;
};

export interface TournamentStats {
  totalMatches: number;
  upcomingMatches: number;
  liveMatches: number;
  completedMatches: number;
  totalPlayers: number;
  totalRegistrations: number;
  soloMatches: number;
  duoMatches: number;
  squadMatches: number;
}

export type ListMatchesParams = {
  type?: ListMatchesType;
  status?: ListMatchesStatus;
};

export type ListMatchesType =
  (typeof ListMatchesType)[keyof typeof ListMatchesType];

export const ListMatchesType = {
  solo: "solo",
  duo: "duo",
  squad: "squad",
} as const;

export type ListMatchesStatus =
  (typeof ListMatchesStatus)[keyof typeof ListMatchesStatus];

export const ListMatchesStatus = {
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;