# lib/api-zod/src/generated/types/adminRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Fire Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { Match } from "./match";
import type { Registration } from "./registration";

export type AdminRegistration = Registration & {
  match: Match;
};